Hello

+++ DefaultSelect.java
@@ -62,7 +62,7 @@ public abstract class DefaultSelect implements Comparable<Integer>{
         //printArray(myArray);
 
         //System.out.println(myArray[Math.floor((myArray.length + 1)/2)]);
-        int index = Math.floor((myArray.length+1)/2);
+        int index = (int)Math.floor((myArray.length+1)/2);
         System.out.println(myArray[index]);
     }
 }
